requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
# chardet==3.0.4
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.47.0
django-cors-headers==3.10.0
django-prometheus==2.1.0
django-rest-auth==0.9.5
# django-utils-six==2.0
djangorestframework==3.12.4
# djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
idna==3.3
oauthlib==3.1.1
prometheus-client==0.9.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
# six==1.11.0
urllib3==1.26.8
uWSGI==2.0.20
whitenoise==5.2.0
